+++ glos-analyzer/src/lib.rs
@@ -1,1 +1,3 @@
+pub mod spectrum;
 
+pub use spectrum::*;